Do not send an empty volume list to delete() during test cleanup

+++ nephoria/testcases/ec2/instances/legacy_instance_test_suite.py
@@ -249,7 +249,8 @@ class LegacyInstanceTestSuite(CliTestRunner):
                     volume.update()
                     if volume.status != 'deleted':
                         delete.append(volume)
-                self.user.ec2.delete_volumes(delete)
+                if delete:
+                    self.user.ec2.delete_volumes(delete)
         except Exception as E:
             self.log.error(red(get_traceback()))
             errors.append(E)
